# Don't allocate a mergeSeries if there is only one series to merge.