~Might help #3715~

Doesn't actually seem to help, but are good to have none-the-less.