@grobie 2018-01-25
Do we have a benchmark showing the impact in terms of allocations? If not, could we add one, so that we can catch regressions in future releases?