@tomwilkie 2018-01-25
> Do we have a benchmark showing the impact in terms of allocations?

Not thats I know of :-(

> If not, could we add one, so that we can catch regressions in future releases?

Sure, will try and get it done today.