@tomwilkie 2018-01-26
Before these changes:
```
$ go test -bench=MergeSeriesSet ./storage/
BenchmarkNoMergeSeriesSet_100_100-4      	500000000	         2.78 ns/op
BenchmarkMergeSeriesSet_1_100_100-4      	200000000	         6.88 ns/op
```

After:
```
$ go test -bench=MergeSeriesSet ./storage/
BenchmarkNoMergeSeriesSet_100_100-4      	1000000000	         2.51 ns/op
BenchmarkMergeSeriesSet_1_100_100-4      	1000000000	         2.89 ns/op
```

Basically, calling MergeSeriesSet on a single SeriesSet is now no extra cost.  I added some extra benchmarks but they need more reasoning first.